backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.2.0
openpyxl==3.1.2
python-multipart==0.0.6
psycopg2-binary>=2.9.10
pyjwt==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator>=2.0.0